backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.2
pydantic[email]==2.9.2
pydantic_settings==2.5.2 
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
transformers==4.44.2
torch==2.4.1
celery==5.4.0
redis==5.0.8
httpx==0.27.2
websockets==13.1
reportlab==4.2.5
python-dotenv==1.0.1
bcrypt==4.1.3
openai>=1.50.0
sentencepiece==0.2.0
deep-translator==1.11.4
langdetect==1.0.9
safetensors

